+++ src/inspect_ai/tool/_tools/_web_browser/_resources/Dockerfile
