+++ src/inspect_ai/tool/_tools/_web_browser/_resources/images/usage_diagram.png
